Articles and Users controller updated

+++ src/articles/articles.controller.ts
@@ -8,15 +8,20 @@ import {
   Delete,
   ParseIntPipe,
   NotFoundException,
+  UseInterceptors,
+  UseGuards,
+  HttpStatus,
 } from '@nestjs/common';
 import { ArticlesService } from './articles.service';
 import { CreateArticleDto } from './dto/create-article.dto';
 import { UpdateArticleDto } from './dto/update-article.dto';
-import { ApiCreatedResponse, ApiOkResponse, ApiTags } from '@nestjs/swagger';
+import { ApiBearerAuth, ApiCreatedResponse, ApiOkResponse, ApiTags } from '@nestjs/swagger';
 import { ArticleEntity } from './entities/article.entity';
-
+import { TransformInterceptor } from 'src/interceptors/transform.interceptor';
+import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
 @Controller('articles')
 @ApiTags('articles')
+@UseInterceptors(TransformInterceptor)
 export class ArticlesController {
   constructor(private readonly articlesService: ArticlesService) {}
 
@@ -29,43 +34,64 @@ export class ArticlesController {
   }
 
   @Get('drafts')
+  @UseGuards(JwtAuthGuard)
+  @ApiBearerAuth()
   @ApiOkResponse({ type: ArticleEntity, isArray: true })
   async findDrafts() {
     const drafts = await this.articlesService.findDrafts();
-    return drafts.map((draft) => new ArticleEntity(draft));
+    const result = drafts.map((draft) => new ArticleEntity(draft));
+    return { statusCode: HttpStatus.OK, message:"Articles Fetched Successfully", result:result};
   }
 
   @Get()
+  @UseGuards(JwtAuthGuard)
+  @ApiBearerAuth()
   @ApiOkResponse({ type: ArticleEntity, isArray: true })
   async findAll() {
     const articles = await this.articlesService.findAll();
-    return articles.map((article) => new ArticleEntity(article));
+    const result = articles.map((article) => new ArticleEntity(article));
+    return { statusCode: HttpStatus.OK, message:"Articles Fetched Successfully", result:result};
   }
 
   @Get(':id')
+  @UseGuards(JwtAuthGuard)
+  @ApiBearerAuth()
   @ApiOkResponse({ type: ArticleEntity })
   async findOne(@Param('id', ParseIntPipe) id: number) {
     const article = await this.articlesService.findOne(id);
     if (!article) {
       throw new NotFoundException(`Article with ${id} does not exist.`);
     }
-    return new ArticleEntity(article);
+    const result = new ArticleEntity(article);
+    return { statusCode: HttpStatus.OK, message:"Article Fetched Successfully", result:result};
   }
 
   @Patch(':id')
+  @UseGuards(JwtAuthGuard)
+  @ApiBearerAuth()
   @ApiOkResponse({ type: ArticleEntity })
   async update(
     @Param('id', ParseIntPipe) id: number,
     @Body() updateArticleDto: UpdateArticleDto,
   ) {
-    return new ArticleEntity(
-      await this.articlesService.update(id, updateArticleDto),
-    );
+    const article = await this.articlesService.findOne(id);
+    if (!article) {
+      throw new NotFoundException(`Article with ${id} does not exist.`);
+    }
+    const result = new ArticleEntity(await this.articlesService.update(id, updateArticleDto));
+    return { statusCode: HttpStatus.OK, message:"Article Updated Successfully", result:result};
   }
 
   @Delete(':id')
+  @UseGuards(JwtAuthGuard)
+  @ApiBearerAuth()
   @ApiOkResponse({ type: ArticleEntity })
   async remove(@Param('id', ParseIntPipe) id: number) {
-    return new ArticleEntity(await this.articlesService.remove(id));
+    const article = await this.articlesService.findOne(id);
+    if (!article) {
+      throw new NotFoundException(`Article with ${id} does not exist.`);
+    }
+    const result = new ArticleEntity(await this.articlesService.remove(id));
+    return { statusCode: HttpStatus.OK, message:"Article Removed Successfully", result:result};
   }
 }
